+++ user-operator/src/test/java/io/strimzi/operator/user/model/KafkaUserModelTest.java
@@ -15,6 +15,7 @@ import io.strimzi.api.kafka.model.user.KafkaUserSpec;
 import io.strimzi.api.kafka.model.user.KafkaUserTlsClientAuthentication;
 import io.strimzi.api.kafka.model.user.KafkaUserTlsExternalClientAuthentication;
 import io.strimzi.certs.CertManager;
+import io.strimzi.certs.OpenSslCertManager;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.Util;
 import io.strimzi.operator.common.model.InvalidResourceException;
@@ -27,6 +28,9 @@ import org.junit.jupiter.api.Test;
 
 import java.nio.charset.StandardCharsets;
 import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneOffset;
+import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.Base64;
 import java.util.HashSet;
@@ -289,6 +293,160 @@ public class KafkaUserModelTest {
 
         // Check owner reference
         checkOwnerReference(model.createOwnerReference(), generatedSecret);
+    }
+
+    @Test
+    public void testGenerateSecretWithRenewalDaysHigherThanValidityDays() {
+        KafkaUser kafkaUserWithValidity = new KafkaUserBuilder(ResourceUtils.createKafkaUserTls(ResourceUtils.NAMESPACE))
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withValidityDays(10)
+                    .withRenewalDays(11)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        KafkaUserModel model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        assertThrows(InvalidResourceException.class,
+            () -> model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, mockCertManager, passwordGenerator, clientsCaCert, clientsCaKey, null, 365, 30, null, Clock.systemUTC()));
+    }
+
+    @Test
+    public void testGenerateSecretWithDefaultRenewalDaysHigherThanValidityDays() {
+        KafkaUser kafkaUserWithValidity = new KafkaUserBuilder(ResourceUtils.createKafkaUserTls(ResourceUtils.NAMESPACE))
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withValidityDays(10)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        KafkaUserModel model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        assertThrows(InvalidResourceException.class,
+            () -> model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, mockCertManager, passwordGenerator, clientsCaCert, clientsCaKey, null, 365, 30, null, Clock.systemUTC()));
+    }
+
+    @Test
+    public void testGenerateSecretWithCurrentValidityExceedingNewValidityPolicy() {
+        int renewalDays = 5;
+        int oldValidityDays = 30;
+        int newValidityDays = 10;
+        Clock currentTime = Clock.systemUTC();
+
+        KafkaUser kafkaUserWithValidity = new KafkaUserBuilder(ResourceUtils.createKafkaUserTls(ResourceUtils.NAMESPACE))
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withValidityDays(oldValidityDays)
+                    .withRenewalDays(renewalDays)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        KafkaUserModel model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, new OpenSslCertManager(currentTime), passwordGenerator, clientsCaCert, clientsCaKey, null, 365, 30, null, currentTime);
+
+        Secret oldCertificate = model.generateSecret();
+
+        // Change validityDays to number that will make the current validity exceeding the new one
+        kafkaUserWithValidity = new KafkaUserBuilder(kafkaUserWithValidity)
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withRenewalDays(renewalDays)
+                    .withValidityDays(newValidityDays)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, new OpenSslCertManager(currentTime), passwordGenerator, clientsCaCert, clientsCaKey, oldCertificate, 365, 30, null, currentTime);
+
+        Secret newCertificate = model.generateSecret();
+
+        assertThat(oldCertificate.equals(newCertificate), is(false));
+    }
+
+    @Test
+    public void testGenerateSecretWithCertificateExpiringUnderNewValidityPolicy() {
+        int renewalDays = 5;
+        int oldValidityDays = 30;
+        int newValidityDays = 20;
+
+        Instant now = Instant.now();
+        Clock issuedClock = Clock.fixed(now.minus(25, ChronoUnit.DAYS), ZoneOffset.UTC);
+        Clock currentTime = Clock.fixed(now, ZoneOffset.UTC);
+
+        KafkaUser kafkaUserWithValidity = new KafkaUserBuilder(ResourceUtils.createKafkaUserTls(ResourceUtils.NAMESPACE))
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withValidityDays(oldValidityDays)
+                    .withRenewalDays(renewalDays)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        KafkaUserModel model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, new OpenSslCertManager(issuedClock), passwordGenerator, clientsCaCert, clientsCaKey, null, 365, 30, null, issuedClock);
+
+        Secret oldCertificate = model.generateSecret();
+
+        // Change validityDays to number that will make the current validity exceeding the new one
+        kafkaUserWithValidity = new KafkaUserBuilder(kafkaUserWithValidity)
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withRenewalDays(renewalDays)
+                    .withValidityDays(newValidityDays)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, new OpenSslCertManager(currentTime), passwordGenerator, clientsCaCert, clientsCaKey, oldCertificate, 365, 30, null, currentTime);
+
+        Secret newCertificate = model.generateSecret();
+
+        assertThat(oldCertificate.equals(newCertificate), is(false));
+    }
+
+    @Test
+    public void testGenerateSecretWithNewValidityPolicy() {
+        int renewalDays = 5;
+        int oldValidityDays = 60;
+        int newValidityDays = 70;
+
+        Instant now = Instant.now();
+        Clock issuedClock = Clock.fixed(now.minus(25, ChronoUnit.DAYS), ZoneOffset.UTC);
+        Clock currentTime = Clock.fixed(now, ZoneOffset.UTC);
+
+        KafkaUser kafkaUserWithValidity = new KafkaUserBuilder(ResourceUtils.createKafkaUserTls(ResourceUtils.NAMESPACE))
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withValidityDays(oldValidityDays)
+                    .withRenewalDays(renewalDays)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        KafkaUserModel model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, new OpenSslCertManager(issuedClock), passwordGenerator, clientsCaCert, clientsCaKey, null, 365, 30, null, issuedClock);
+
+        Secret oldCertificate = model.generateSecret();
+
+        // Change validityDays to number that will make the current validity exceeding the new one
+        kafkaUserWithValidity = new KafkaUserBuilder(kafkaUserWithValidity)
+            .editSpec()
+                .withNewKafkaUserTlsClientAuthentication()
+                    .withRenewalDays(renewalDays)
+                    .withValidityDays(newValidityDays)
+                .endKafkaUserTlsClientAuthentication()
+            .endSpec()
+            .build();
+
+        model = KafkaUserModel.fromCrd(kafkaUserWithValidity, UserOperatorConfig.SECRET_PREFIX.defaultValue(), Boolean.parseBoolean(UserOperatorConfig.ACLS_ADMIN_API_SUPPORTED.defaultValue()));
+        model.maybeGenerateCertificates(Reconciliation.DUMMY_RECONCILIATION, new OpenSslCertManager(currentTime), passwordGenerator, clientsCaCert, clientsCaKey, oldCertificate, 365, 30, null, currentTime);
+
+        Secret newCertificate = model.generateSecret();
+
+        assertThat(oldCertificate.equals(newCertificate), is(true));
     }
 
     @Test